+++ package.json
@@ -16,40 +16,40 @@
     "copy_readme": "copyfiles ./README.md dist/translate"
   },
   "dependencies": {
-    "@angular/animations": "^14.0.3",
-    "@angular/cdk": "^13.1.2",
-    "@angular/cli": "^14.0.3",
-    "@angular/common": "^14.0.3",
-    "@angular/compiler": "^14.0.3",
-    "@angular/compiler-cli": "^14.0.3",
-    "@angular/core": "^14.0.3",
-    "@angular/forms": "^14.0.3",
-    "@angular/language-service": "^14.0.3",
-    "@angular/material": "^13.1.2",
-    "@angular/platform-browser": "^14.0.3",
-    "@angular/platform-browser-dynamic": "^14.0.3",
-    "@angular/router": "^14.0.3",
+    "@angular/animations": "^14.2.8",
+    "@angular/cdk": "^14.2.6",
+    "@angular/cli": "^14.2.7",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/compiler-cli": "^14.2.8",
+    "@angular/core": "^14.2.8",
+    "@angular/forms": "^14.2.8",
+    "@angular/language-service": "^14.2.8",
+    "@angular/material": "^14.2.6",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/router": "^14.2.8",
     "rxjs": "^7.3.0",
-    "simply-translate": "^0.20.0-beta.18",
+    "simply-translate": "^0.20.0",
     "tslib": "^2.3.1",
     "typescript": "4.7.4",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.0.3",
-    "@types/jasmine": "~3.10.3",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@types/jasmine": "~4.3.0",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "~17.0.8",
+    "@types/node": "~18.11.7",
     "copyfiles": "^2.4.1",
-    "core-js": "^3.23.3",
-    "jasmine-core": "~4.0.0",
+    "core-js": "^3.26.0",
+    "jasmine-core": "~4.4.0",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.11",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
     "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^14.0.2",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "tslint": "~6.1.3"
   }
