+++ projects/translate/package.json
@@ -1,10 +1,10 @@
 {
   "name": "simply-translate-angular",
-  "version": "0.20.0-beta.18",
+  "version": "0.20.0",
   "dependencies": {
-    "tslib": "^2.0.0",
-    "simply-translate": "0.20.0-beta.18",
-    "rxjs": "^7.3.0"
+    "tslib": "^2.4.0",
+    "simply-translate": "0.20.0",
+    "rxjs": "^7.5.7"
   },
   "peerDependencies": {
     "@angular/common": "*",
@@ -23,8 +23,8 @@
     "localize",
     "localization"
   ],
-  "author": "Oleg A.V.",
-  "license": "ISC",
+  "author": "Oleg Volovikov",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/oleg-wx/translate-angular/issues"
   },
