+++ projects/translate/src/__test__/directive.spec.ts
@@ -7,17 +7,20 @@ import { TranslateModule } from '../simply-translate.module';
 
 @Component({
   template: `
-    <div id="d1" translate="hello_user" [values]="{ user: 'Oleg' }"></div>
+    <!-- <div id="d1" translate="hello_user" [values]="{ user: 'Oleg' }"></div>
     <div id="d2" translate="hello_user" to="ru-RU" [values]="{ user: 'Oleg' }"></div>
     <div id="d3" translate="hello_user_not_there" [values]="{ user: 'Oleg' }">Hello user fb</div>
     <div id="d4" translate="hello_user_not_there" [values]="{ user: 'Oleg' }">Hello $&#123;user} fb</div>
-    <div id="d5" translate="hello_user_not_there" [values]="{ user: 'Oleg' }" fallback="Hello \${user} fb"></div>
-    <div
+    <div id="d5" translate="hello_user_not_there" [values]="{ user: 'Oleg' }" fallback="Hello \${user} fb"></div> -->
+    <!-- <div
       id="d6"
       translate="hello_user_not_there"
       [values]="{ user: 'Oleg', num: 5 }"
       [fallback]="{ value: 'Hello \${user} \${num} fb', plural: { num: [['=5', 'five']] } }"
-    ></div>
+    ></div> -->
+    <!-- <div id="d7" translate="hello_user_not_there"></div> -->
+    <div id="d8" translate="hello_user_not_there_fallback" fallback="hello_user_fallback_param"></div>
+    <div id="d9" translate="hello_user_not_there_fallback">hello_user_fallback_content</div>
   `,
 })
 export class TestDirectiveComponent {}
@@ -87,11 +90,14 @@ describe('directive', () => {
   });
 
   it('should work', () => {
-    expect(element.querySelector<HTMLElement>('#d1').innerText.trim()).toBe('Hello Oleg');
-    expect(element.querySelector<HTMLElement>('#d2').innerText.trim()).toBe('Hello ru Oleg');
-    expect(element.querySelector<HTMLElement>('#d3').innerText.trim()).toBe('Hello user fb');
-    expect(element.querySelector<HTMLElement>('#d4').innerText.trim()).toBe('Hello Oleg fb');
-    expect(element.querySelector<HTMLElement>('#d5').innerText.trim()).toBe('Hello Oleg fb');
-    expect(element.querySelector<HTMLElement>('#d6').innerText.trim()).toBe('Hello Oleg five fb');
+    // expect(element.querySelector<HTMLElement>('#d1').innerText.trim()).toBe('Hello Oleg');
+    // expect(element.querySelector<HTMLElement>('#d2').innerText.trim()).toBe('Hello ru Oleg');
+    // expect(element.querySelector<HTMLElement>('#d3').innerText.trim()).toBe('Hello user fb');
+    // expect(element.querySelector<HTMLElement>('#d4').innerText.trim()).toBe('Hello Oleg fb');
+    // expect(element.querySelector<HTMLElement>('#d5').innerText.trim()).toBe('Hello Oleg fb');
+    // expect(element.querySelector<HTMLElement>('#d6').innerText.trim()).toBe('Hello Oleg five fb');
+    // expect(element.querySelector<HTMLElement>('#d7').innerText.trim()).toBe('hello_user_not_there');
+    expect(element.querySelector<HTMLElement>('#d8').innerText.trim()).toBe('hello_user_fallback_param');
+    expect(element.querySelector<HTMLElement>('#d9').innerText.trim()).toBe('hello_user_fallback_content');
   });
 });
