+++ projects/translate/src/public_api.ts
@@ -3,7 +3,7 @@ export * from './translate/translate.pipe';
 export * from './translate/translate.resolver';
 export * from './translate/translate.service';
 export * from './simply-translate.module';
-export { TranslateChildConfig } from './translate/translate-child-config';
+export { TranslateChildConfig } from './translate/translate.service';
 export {
   Dictionary,
   Dictionaries,
@@ -12,5 +12,5 @@ export {
   TranslateDynamicProps,
   Pipeline,
   SimplePipeline,
-  SimpleDefaultPipeline
+  SimpleDefaultPipeline,
 } from 'simply-translate';
